Methods that don't belong on Scalars should stay off Scalar.
For example `.S` and `__getitem__`.
